Finish Recipe page + adding exportprice column in ExportRecipeItem model and change amount column in Recipe model from Int to Float

+++ api/models/Recipe.js
@@ -14,7 +14,7 @@ module.exports = {
 
         ingredient: { model: 'ingredient', required: true },
 
-        amount: { type: 'integer', required: true },
+        amount: { type: 'float', required: true },
 
   }
 };
